requirements.txt
opencontainers
python-dateutil
humanize